DC_EDA Presentation
I went ahead and dropped in some of the bullets for A-D into slides for our presentation. Feel free to drop in your univariate forecasts if you're feeling confident in them (i'm not feeling great about the ASE scores i got, but i figure that's the nature of the data and the fact we aren't using any other variables). Otherwise, I can drop in what I found and we can record on that.